Create counter component
